improve email input to fit center ui
With the Microsoft [Centered UI Web Theme][1], the layout around the
email input concerning TOTP enrollment could be improved. For end-users
it's not clear what they should fill in, or where they should fill it
in.

This tries to improve the layout to work with both the centered ui and
the original ui theme.

[1]:https://github.com/Microsoft/adfsWebCustomization/tree/master/centeredUi