Allow providing full e-mail address
The code currently assumes the user provides only the user part of the
e-mail address and unconditionally appends the known domain to the user
provided data.

This results in an error message when the user specifies the full e-mail
address.

Only append the domain when no '@' is present in the user provided data.